Refactor: Extract userReducer to separate file
