+++ client/src/context/UserContext.jsx
@@ -8,34 +8,9 @@ import {
 import { defaultUser } from "../data/defaultUser";
 import { fixUserSkills } from "../util/fixUserSkills";
 import useFetch from "../hooks/useFetch";
+import { userReducer } from "../reducers/userReducer";
 
-function userReducer(state, action) {
-  switch (action.type) {
-    case "REGISTER":
-    case "LOGIN":
-      return { ...defaultUser, ...action.payload };
-    case "UPDATE_USER": {
-      return { ...state, ...action.payload };
-    }
-    case "LOGOUT":
-      return defaultUser;
-    case "SET_SKILLS": {
-      return { ...state, skills: action.payload };
-    }
-    case "TOGGLE_FAVORITE": {
-      const job = action.payload;
-      const jobId = job?.id;
-      const prevFavorites = state.favorites;
-      const exists = prevFavorites.some((fav) => fav.id === jobId);
-      const newFavorites = exists
-        ? prevFavorites.filter((fav) => fav.id !== jobId)
-        : [...prevFavorites, job];
-      return { ...state, favorites: newFavorites };
-    }
-    default:
-      return state;
-  }
-}
+
 
 const UserContext = createContext();
